requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-core
langgraph
huggingface-hub>=0.25.0
sentence-transformers>=2.7.0
faiss-cpu
fastapi>=0.111.0
uvicorn>=0.29.0
streamlit>=1.35.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv
pyngrok
nest-asyncio
httpx>=0.27.0
requests
openai
